recipe-upload/requirements.txt
python-dotenv
selenium
supabase
requests
sentence_transformers
pandas
datasets